+++ Example/Tests/PlayerEventsTest.swift
@@ -366,8 +366,7 @@ class PlayerEventsTest: QuickSpec {
                             haveBeenCalled(
                                 withArgs: [
                                     "adPlayer": "\(AdPlayer.ADPLAYER_CONTENT)",
-                                    "adType": "\(AdTechnology.CLIENT_SIDE)",
-                                    "adBreakInfo": "\(AdPosition.ADPOSITION_PREROLL.rawValue)"
+                                    "adType": "\(AdTechnology.CLIENT_SIDE)"
                                 ]
                             )
                         )